[torch.compile x autograd.Function] More test cases
I pulled a bunch of autograd.Function from test_autograd.py and added a
smoke test for them. Ideally we would actually run test_autograd.py as a
part of the Dynamo test suite, but we have excluded it due to there
being too many errors and I don't have time to figure that out at the
moment.

ghstack-source-id: 39aadb02bf325fe3a1be0ad9e9e48904b33bdc30
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107467