[Quantization] Add metadata porting for nodes added by quantization
Summary:
This diff adds adding metadata to q-dq nodes by inferring the
quatization intent from node annotations. Annotations on the node are
way for user to specify how a node or subgraph is supposed to be
quantized. We continue to use that information to copy metadata on Q/DQ
node from appropriate nodes.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 82cd0f45ef7b265b4207e1f5b82da0f86d196152
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107107