[WIP] Add CPU-side profiler event for triton kernels w/ python wrapper
If you want to view the kernel name for a kernel in a trace (e.g. to reference the triton kernel implementation in the python wrapper code / TORCH_COMPILE_DEBUG logs), but you don't want to turn on unique_kernel_names (because unique_kernel_names=False will maximize caching), then this provides a solution.

ghstack-source-id: db57b2badb253733a242d1207cadc5d822ca1123
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106351