[allocator] register oom observers on every device
This change is to match the behavior of _record_memory_history which was
recently changed to enable history recording on all devices rather than
the current one. It prevents confusing situations where the observer
was registered before the device was set for the training run.

It also ensures the Python bindings initialize cuda so that the allocators
have been created before registering the callbacks.

ghstack-source-id: 9a648bba445791c952e283fb82fcd20690a2e5f2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107399